[Ref] mpt/crypto/jwk.hpp: Silence Clang warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@19480 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ src/mpt/crypto/jwk.hpp
@@ -332,7 +332,7 @@ public:
 			if (result == 0x00000000 /*STATUS_SUCCESS*/) {
 				return;
 			}
-			if (result == 0xC000A000 /*STATUS_INVALID_SIGNATURE*/) {
+			if (static_cast<DWORD>(result) == 0xC000A000 /*STATUS_INVALID_SIGNATURE*/) {
 				throw signature_verification_failed();
 			}
 			CheckNTSTATUS(result, "BCryptVerifySignature");
